See commit message for fixes, changes, and new features. Thanks!
Changes
- Updated mex files for tree_v1 and tree_v8

Features
- Added silent running feature to INSTALL_PMA_project.m
- Added tree_v10_mex option (seems faster in some cases but more testing is needed)

+++ src/tree/v10/PMA_EnumerateAlg10.m
@@ -136,7 +136,7 @@ for iter = 1:Ne
     if dispflag > 2 % very verbose
         fprintf('---\n')    
         fprintf('Iteration: %2i\n',iter)
-        fprintf('       Current Queue Length: %8d\n',length(Queue))
+        fprintf('       Current Queue Length: %8d\n',int64(length(Queue)))
     end
     %----------------------------------------------------------------------
     
@@ -156,7 +156,7 @@ for iter = 1:Ne
         
         % print
         if dispflag > 2 % very verbose
-            fprintf('Removed Graphs (Simple ISO): %8d\n',NQueue-length(Queue))
+            fprintf('Removed Graphs (Simple ISO): %8d\n',NQueue-int64(length(Queue)))
         end
     end
     %----------------------------------------------------------------------
